server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
firebase-admin==6.5.0
reportlab==4.2.2
python-dotenv==1.0.1
joblib==1.4.2
numpy==2.1.1
scikit-learn==1.5.2
google-genai==1.33.0
xgboost==3.2.0
